[runtime-common] extract constants module
