+++ polkadot-parachains/common/src/lib.rs
@@ -13,70 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//!
+//! Common definitions for runtimes. It contains both: definitions by encointer, and
+//! definitions from the statemine runtime.
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod deal_with_fees;
 pub mod weights;
 
-// below is copied from cumulus/parachains/runtimes/assets/statemine/constants
-
-pub mod currency {
-	use kusama_runtime_constants as constants;
-	use polkadot_core_primitives::Balance;
-
-	/// The existential deposit. Set to 1/10 of its parent Relay Chain.
-	pub const EXISTENTIAL_DEPOSIT: Balance = constants::currency::EXISTENTIAL_DEPOSIT / 10;
-
-	pub const UNITS: Balance = constants::currency::UNITS;
-	pub const CENTS: Balance = constants::currency::CENTS;
-	pub const GRAND: Balance = constants::currency::GRAND;
-	pub const MILLICENTS: Balance = constants::currency::MILLICENTS;
-
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		// map to 1/10 of what the kusama relay chain charges (v9020)
-		constants::currency::deposit(items, bytes) / 10
-	}
-}
-
-/// Fee-related.
-pub mod fee {
-	use frame_support::weights::{
-		constants::ExtrinsicBaseWeight, WeightToFeeCoefficient, WeightToFeeCoefficients,
-		WeightToFeePolynomial,
-	};
-	use polkadot_core_primitives::Balance;
-	use smallvec::smallvec;
-	pub use sp_runtime::Perbill;
-
-	/// The block saturation level. Fees will be updates based on this value.
-	pub const TARGET_BLOCK_FULLNESS: Perbill = Perbill::from_percent(25);
-
-	/// Handles converting a weight scalar to a fee value, based on the scale and granularity of the
-	/// node's balance type.
-	///
-	/// This should typically create a mapping between the following ranges:
-	///   - [0, MAXIMUM_BLOCK_WEIGHT]
-	///   - [Balance::min, Balance::max]
-	///
-	/// Yet, it can be used for any other sort of change to weight-fee. Some examples being:
-	///   - Setting it to `0` will essentially disable the weight fee.
-	///   - Setting it to `1` will cause the literal `#[weight = x]` values to be charged.
-	pub struct WeightToFee;
-	impl WeightToFeePolynomial for WeightToFee {
-		type Balance = Balance;
-		fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-			// in Kusama, extrinsic base weight (smallest non-zero weight) is mapped to 1/10 CENT:
-			// in Statemine, we map to 1/10 of that, or 1/100 CENT
-			let p = super::currency::CENTS;
-			let q = 100 * Balance::from(ExtrinsicBaseWeight::get());
-			smallvec![WeightToFeeCoefficient {
-				degree: 1,
-				negative: false,
-				coeff_frac: Perbill::from_rational(p % q, q),
-				coeff_integer: p / q,
-			}]
-		}
-	}
-}
+// copied from cumulus/parachains/runtimes/assets/statemine/constants
+mod constants;
+pub use constants::{currency, fee};
